Implement additional parameter for polling latest version from specified tag inside maven-metadata.xml
The name of the tag can be configured in the GoCD UI.
If there none tag specified or the specified tag can not be found in maven-metadata.xm, the poller behaves as before: look for all available versions and take the greatest one
If one enters a tag name (e.g. release) the version inside this xml tag of maven-metadata.xml is taken.